+++ packages/core-js/modules/esnext.array.from-async.js
@@ -1,0 +1,8 @@
+var $ = require('../internals/export');
+var fromAsync = require('../internals/array-from-async');
+
+// `Array.fromAsync` method
+// https://github.com/tc39/proposal-array-from-async
+$({ target: 'Array', stat: true }, {
+  fromAsync: fromAsync
+});