+++ packages/core-js/proposals/array-from-async.js
@@ -1,0 +1,3 @@
+// https://github.com/tc39/proposal-array-from-async
+require('../modules/esnext.array.from-async');
+require('../modules/esnext.typed-array.from-async');