+++ tests/compat/tests.js
@@ -1198,6 +1198,9 @@ GLOBAL.tests = {
       && set.add({}) == set
       && set[Symbol.toStringTag];
   }],
+  'esnext.array.from-async': function () {
+    return Array.fromAsync;
+  },
   'esnext.array.filter-reject': function () {
     return [].filterReject;
   },
@@ -1469,6 +1472,9 @@ GLOBAL.tests = {
   },
   'esnext.symbol.observable': function () {
     return Symbol.observable;
+  },
+  'esnext.typed-array.from-async': function () {
+    return Int8Array.fromAsync;
   },
   'esnext.typed-array.filter-reject': function () {
     return Int8Array.prototype.filterReject;